Great update to index.py
Basically is this creates both "Images" and "PDF" folders if they don't already exist, to make life more easier. It checks if there are any .JPG or .PNG files in the "Images" folder, printing a warning message and stopping the execution if none are found. but you can add more extensions modifying the 17 line... Finally I made it save the date and time the PDF was create for the sake of organization,..